Move precise-code-intel-worker out of enterprise
Not required anymore, just a git mv and some path changes.
