+++ cmd/precise-code-intel-worker/image_test.yaml
