+++ cmd/precise-code-intel-worker/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/precise-code-intel-worker/shared"
+	"github.com/sourcegraph/sourcegraph/cmd/precise-code-intel-worker/shared"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/sourcegraph/enterprisecmd"
 	"github.com/sourcegraph/sourcegraph/internal/sanitycheck"
 )
