+++ cmd/precise-code-intel-worker/shared/service.go
