+++ enterprise/cmd/sourcegraph/main.go
@@ -11,12 +11,12 @@ import (
 
 	blobstore_shared "github.com/sourcegraph/sourcegraph/cmd/blobstore/shared"
 	githubproxy_shared "github.com/sourcegraph/sourcegraph/cmd/github-proxy/shared"
+	precise_code_intel_worker_shared "github.com/sourcegraph/sourcegraph/cmd/precise-code-intel-worker/shared"
 	searcher_shared "github.com/sourcegraph/sourcegraph/cmd/searcher/shared"
 	embeddings_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/embeddings/shared"
 	executor_singlebinary "github.com/sourcegraph/sourcegraph/enterprise/cmd/executor/singlebinary"
 	frontend_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/shared"
 	gitserver_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/gitserver/shared"
-	precise_code_intel_worker_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/precise-code-intel-worker/shared"
 	repoupdater_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/repo-updater/shared"
 	symbols_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/symbols/shared"
 	worker_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/shared"
